requirements.txt
langchain==0.3.3
langchain-google-genai==2.0.1
python-dotenv==1.0.1
pymongo
pillow